requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly
fastapi
uvicorn
streamlit
mlflow
pydantic
python-dotenv
pyyaml
joblib
shap
requests
